ultimazione dao service e controller per tutto quello che rigurda la gestione dei pagamenti e progressi per quanto riguarda quella della gestione amicizia

+++ MoVita_BackEnd/src/main/java/org/example/movita_backend/model/Payment.java
@@ -37,7 +37,7 @@ public class Payment {
         this.titolo = titolo;
     }
 
-    public int getAmmontare() {
+    public long getAmmontare() {
         return ammontare;
     }
